uucore(fs): make display_permissions_unix cross-platform (#11538)
Remove the #[cfg(unix)] gate from display_permissions_unix and its
helper get_file_display. Change the parameter from libc::mode_t to u32
(Metadata::mode() already returns u32, so no cast or information loss).

Add pub mod mode to uucore::fs with portable u32 constants for all POSIX
S_I* values (S_IFMT, S_IFSOCK, S_IFLNK, S_IFREG, S_IFBLK, S_IFDIR,
S_IFCHR, S_IFIFO, S_ISUID, S_ISGID, S_ISVTX, and all nine rwx bits).
Callers can import uucore::fs::mode::* instead of libc, without
requiring a Unix target.

Replace the has!() macro calls in display_permissions_unix with explicit
bitwise & != 0 checks, which work without a type-specific macro.

Update chmod and mv to drop the libc::mode_t import and the as mode_t
casts at each call site. Update is_stdin_directory to use mode::S_IFMT
and mode::S_IFDIR instead of the libc imports.

Remove #[cfg(unix)] from test_display_permissions and test_get_file_display
so both tests compile and run on all platforms.

+++ src/uu/cp/src/cp.rs
@@ -1620,7 +1620,7 @@ fn file_mode_for_interactive_overwrite(
 
                         Some((
                             format!("{mode_without_leading_digits:04o}"),
-                            uucore::fs::display_permissions_unix(mode, false),
+                            uucore::fs::display_permissions_unix(mode as u32, false),
                         ))
                     }
                 }
